% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{folder}
\alias{folder}
\title{Folder Entity}
\usage{
folder(
  name = "",
  description = "",
  id = "",
  etag = "",
  createdOn = "",
  modifiedOn = "",
  createdBy = "",
  modifiedBy = "",
  parentId = "",
  concreteType = ""
)
}
\arguments{
\item{parentId}{}
}
\value{

}
\description{
Folder Entity
}
