% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{get_entity}
\alias{get_entity}
\title{Get entity}
\usage{
get_entity(id_or_ent)
}
\arguments{
\item{id_or_ent}{Synapse Id or Entity object}
}
\value{
Entity
}
\description{
Get entity
}
\examples{
get_entity("syn...")
}
