% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Synapse login}
\usage{
login(authtoken = NULL)
}
\arguments{
\item{authtoken}{Synapse PAT}
}
\value{
None
}
\description{
Synapse login
}
\examples{
login(authtoken="...")
login()
}
