% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{new_entity}
\alias{new_entity}
\title{Entity object}
\usage{
new_entity(
  x,
  name = character(),
  description = "",
  id = "",
  etag = "",
  createdOn = "",
  modifiedOn = "",
  createdBy = "",
  modifiedBy = "",
  parentId = "",
  concreteType = ""
)
}
\arguments{
\item{name}{Name of entity}

\item{description}{Description of entity}

\item{id}{Synapse Id of entity}

\item{etag}{Etag of entity}

\item{createdOn}{Created on date}

\item{modifiedOn}{Modified on date}

\item{createdBy}{Created by date}

\item{modifiedBy}{modified by date}

\item{parentId}{Synapse Id of the container the entity is contained in.}

\item{concreteType}{Concrete Type}
}
\value{
Entity Class
}
\description{
Entity object
}
\examples{
new_entity("new_ent", name="entity_name", id="syn...")
}
