% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_rest_delete.R
\name{rest_DELETE}
\alias{rest_DELETE}
\title{Synapse rest DELETE command}
\usage{
rest_DELETE(path)
}
\arguments{
\item{path}{Path to synapse endpoint}
}
\value{
Synapse API class
}
\description{
Synapse rest DELETE command
}
\examples{
rest_DELETE("entity/syn26896636")
}
