% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_rest_post.R
\name{rest_POST}
\alias{rest_POST}
\title{Synapse rest POST command}
\usage{
rest_POST(path, body)
}
\arguments{
\item{path}{Path to synapse endpoint}

\item{body}{Request Body}
}
\value{
Synapse API class
}
\description{
Synapse rest POST command
}
\examples{
rest_POST("entity", body = list(
  name = "test-me-now",
  parentId = "syn4990358",
  concreteType = "org.sagebionetworks.repo.model.Folder"
))
}
