% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_rest_put.R
\name{rest_PUT}
\alias{rest_PUT}
\title{Synapse rest PUT command}
\usage{
rest_PUT(path, body)
}
\arguments{
\item{path}{Path to synapse endpoint}

\item{body}{Request Body}
}
\value{
Synapse API class
}
\description{
Synapse rest PUT command
}
\examples{
rest_PUT("entity/syn26896636", body=list(name = "change-my-name"))
}
