% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Log into Synapse via personal access token}
\usage{
login(authtoken = NULL)
}
\arguments{
\item{authtoken}{Synapse PAT}
}
\value{
None
}
\description{
Log into Synapse via personal access token
}
\examples{
login(authtoken="...")
login()
}
