% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_rest_get.R
\name{rest_GET}
\alias{rest_GET}
\title{Synapse REST GET command}
\usage{
rest_GET(path)
}
\arguments{
\item{path}{}
}
\value{
S3 object that is the same as the concreteType returned from Synapse.
}
\description{
Synapse REST GET command
}
\examples{
rest_GET("entity/syn7222066")
}
